requirements.txt
flask==3.0.3
pillow==10.4.0
gunicorn==23.0.0
